move debug primitive generation code to separate file
